#12 Move errors.Close function to its own file
