+++ errors/senders.go
@@ -54,10 +54,3 @@ func Make[E Senders]() (werrs E, rerrs []<-chan error) {
 
 	return
 }
-
-// Close takes slice of error-writing channels and closes its every element.
-func Close(werrs []chan<- error) {
-	for _, ch := range werrs {
-		close(ch)
-	}
-}
